+++ es/tx/builder/helpers.js
@@ -1,16 +1,25 @@
 import * as R from 'ramda'
+import BigNumber from 'bignumber.js'
+
 import {
   assertedType,
   decodeBase58Check,
   decodeBase64Check,
   encodeBase58Check, encodeBase64Check,
   hash,
-  nameId,
+  nameId as nameHash,
   salt
 } from '../../utils/crypto'
 import { toBytes } from '../../utils/bytes'
-import { ID_TAG_PREFIX, PREFIX_ID_TAG, AENS_NAME_DOMAINS } from './schema'
-import { BigNumber } from 'bignumber.js'
+import {
+  ID_TAG_PREFIX,
+  PREFIX_ID_TAG,
+  AENS_NAME_DOMAINS,
+  NAME_BID_RANGES,
+  NAME_BID_MAX_LENGTH,
+  NAME_FEE, NAME_FEE_BID_INCREMENT, NAME_BID_TIMEOUTS
+} from './schema'
+import { ceil } from '../../utils/bignumber'
 
 /**
  * JavaScript-based Transaction builder helper function's
@@ -80,19 +89,16 @@ export function formatSalt (salt) {
 }
 
 /**
- * Generate the commitment hash by hashing the formatted salt and
- * name, base 58 encoding the result and prepending 'cm_'
- *
+ * Encode a domain name
+ * @function
  * @alias module:@aeternity/aepp-sdk/es/tx/builder/helpers
- * @function prelimaCommitmentHash
- * @category async
- * @rtype (name: String, salt?: String) => hash: Promise[String]
- * @param {String} name - Name to be registered
- * @param {Number} salt Random salt
- * @return {String} Commitment hash
+ * @param {String} name Name to encode
+ * @return {String} `nm_` prefixed encoded domain name
  */
-export async function prelimaCommitmentHash (name, salt = createSalt()) {
-  return `cm_${encodeBase58Check(hash(Buffer.concat([nameId(name), formatSalt(salt)])))}`
+export function produceNameId (name) {
+  const namespace = R.last(name.split('.'))
+  if (namespace === 'chain') return encode(hash(name.toLowerCase()), 'nm')
+  return encode(nameHash(name), 'nm')
 }
 
 /**
@@ -107,8 +113,10 @@ export async function prelimaCommitmentHash (name, salt = createSalt()) {
  * @param {Number} salt Random salt
  * @return {String} Commitment hash
  */
-export async function commitmentHash (name, salt = createSalt()) {
-  return `cm_${encodeBase58Check(hash(Buffer.concat([Buffer.from(name), formatSalt(salt)])))}`
+export function commitmentHash (name, salt = createSalt()) {
+  const namespace = R.last(name.split('.'))
+  if (namespace === 'chain') return `cm_${encodeBase58Check(hash(Buffer.concat([Buffer.from(name), formatSalt(salt)])))}`
+  return `cm_${encodeBase58Check(hash(Buffer.concat([nameHash(name.toLowerCase()), formatSalt(salt)])))}`
 }
 
 /**
@@ -225,7 +233,7 @@ export function readPointers (pointers) {
 /**
  * Is name valid
  * @function
- * @alias module:@aeternity/aepp-sdk/es/ae/aens
+ * @alias module:@aeternity/aepp-sdk/es/tx/builder/helpers
  * @param {string} name
  * @param {boolean} [throwError=true] Throw error on invalid
  * @return Boolean
@@ -238,6 +246,62 @@ export function isNameValid (name, throwError = true) {
     return false
   }
   return true
+}
+
+/**
+ * What kind of a hash is this? If it begins with 'ak_' it is an
+ * account key, if with 'ok_' it's an oracle key.
+ *
+ * @param s - the hash.
+ * returns the type, or throws an exception if type not found.
+ */
+export function classify (s) {
+  const keys = {
+    ak: 'account_pubkey',
+    ok: 'oracle_pubkey',
+    ct: 'contract_pubkey',
+    ch: 'channel'
+  }
+
+  if (!s.match(/^[a-z]{2}_.+/)) {
+    throw Error('Not a valid hash')
+  }
+
+  const klass = s.substr(0, 2)
+  if (klass in keys) {
+    return keys[klass]
+  } else {
+    throw Error(`Unknown class ${klass}`)
+  }
+}
+
+/**
+ * Get the minimum name fee for a domain
+ * @function
+ * @alias module:@aeternity/aepp-sdk/es/tx/builder/helpers
+ * @param {string} domain the domain name to get the fee for
+ * @return String the minimum fee for the domain auction
+ */
+export function getMinimumNameFee (domain) {
+  const nameLength = domain.replace('.aet', '').length
+  return NAME_BID_RANGES[nameLength >= NAME_BID_MAX_LENGTH ? NAME_BID_MAX_LENGTH : nameLength]
+}
+
+export function computeBidFee (domain, startFee = NAME_FEE, increment = NAME_FEE_BID_INCREMENT) {
+  if (!(Number(increment) === increment && increment % 1 !== 0)) throw new Error(`Increment must be float. Current increment ${increment}`)
+  if (increment < NAME_FEE_BID_INCREMENT) throw new Error(`minimum increment percentage is ${NAME_FEE_BID_INCREMENT}`)
+  return ceil(
+    BigNumber(BigNumber(startFee).eq(NAME_FEE) ? getMinimumNameFee(domain) : startFee).times(BigNumber(NAME_FEE_BID_INCREMENT).plus(1))
+  )
+}
+
+export function computeAuctionEndBlock (domain, claimHeight) {
+  return R.cond([
+    [R.lt(4), R.always(NAME_BID_TIMEOUTS[1] + claimHeight)],
+    [R.lt(8), R.always(NAME_BID_TIMEOUTS[4] + claimHeight)],
+    [R.lte(NAME_BID_MAX_LENGTH), R.always(NAME_BID_TIMEOUTS[8] + claimHeight)],
+    [R.T, R.always(claimHeight)]
+  ])(domain.replace('.aet', '').length)
 }
 
 export default {
@@ -255,5 +319,7 @@ export default {
   oracleQueryId,
   createSalt,
   buildHash,
-  isNameValid
+  isNameValid,
+  produceNameId,
+  classify
 }